+++ HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 2.3.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fixed broken update step
+  [obct537]
 
 
 2.3.0 (2018-11-27)