+++ HISTORY.rst
@@ -6,7 +6,13 @@ Changelog
 
 New:
 
-- Upgrade to latest collective.elasticsearch
+- Upgrade to latest collective.elasticsearch.
+  New versions include,
+  `collective.celery=1.1.1`
+  `collective.elasticsearch=2.0.2`
+  `celery=4.2.1`
+  `billiard = 3.5.0.4`
+  `kombu = 4.2.1`
   [vangheem]
 
 - Fixed broken update step