+++ HISTORY.rst
@@ -17,6 +17,12 @@ Fixes:
 - Handle error when image tile referenced image is not found
   [vangheem]
 
+- Fix use of celery with always eager setting and some tasks
+  [vangheem]
+
+- Work with ffmpeg as well as avconv
+  [vangheem]
+
 - Make content listing tile persistent. This fixes issues with saving
   data to content listing tile.
   [vangheem]