+++ HISTORY.rst
@@ -6,6 +6,9 @@ Changelog
 
 New:
 
+- Add import subscribers form for announcements panel
+  [CorySanin]
+
 - Add new recurrances indexer
   [tkimnguyen]
 
@@ -72,9 +75,6 @@ Fixes:
 ------------------
 
 New:
-
-- Add import subscribers form for announcements panel
-  [CorySanin]
 
 - Add Mamoto support (CastleCMS API control panel settings, Twitter & Facebook share counting and Analytics display);
   remove EOL'd Facebook and LinkedIn API social counts