+++ HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 2.3.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix upgrade of `castle.slot_tiles` when it's been set to None
+  [vangheem]
 
 
 2.3.3 (2018-12-10)